# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: CampusHub API
  version: 1.0.0
  description: Canonical HTTP API contract for the CampusHub backend foundation.
servers:
  - url: /
paths:
...
  /api/v1/email-codes:
    post:
      operationId: sendEmailCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, scene]
              properties:
                email: {type: string, format: email}
                scene: {type: string, enum: [register, forgot_password, logoff]}
      responses: {'200': {$ref: '#/components/responses/Success'}, '429': {$ref: '#/components/responses/Error'}}
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RefreshRequest'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '401': {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  schemas:
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties: {refresh_token: {type: string}}
...